Move generateID function
So that different components can also use this function
